+++ src/pages/Courses/CoursesPage.tsx
@@ -1,15 +1,16 @@
 import React from 'react';
 import { Footer, Navbar } from '../../components';
-import { Main } from './sections/Main';
-import { About } from './sections/About';
-import { Principles } from '../commonSections/Principles';
-import Courses from './sections/Courses/Courses';
-import StudyPath from './sections/StudyPath/StudyPath';
-import General from './sections/General/General';
-import Mentors from './sections/Mentors/Mentors';
-import Mentoring from './sections/Mentoring/Mentoring';
-import Requirements from './sections/Requirements/Requirements';
-import { Partnered } from '../commonSections/Partnered';
+import { Principles, Partnered } from '../commonSections';
+import {
+  About,
+  Main,
+  Courses,
+  StudyPath,
+  General,
+  Mentors,
+  Mentoring,
+  Requirements
+} from './sections';
 
 export const CoursesPage: React.FC = () => (
   <div className="App">
