+++ src/pages/Courses/sections/Courses/index.tsx
@@ -1,1 +1,1 @@
-export * from '../Courses';
+export * from './Courses';
