+++ src/pages/Courses/sections/index.tsx
@@ -1,0 +1,8 @@
+export * from './About';
+export * from './General';
+export * from './Main';
+export * from './Mentoring';
+export * from './Mentors';
+export * from './Requirements';
+export * from './StudyPath';
+export * from './Courses';
